requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-filter==2.4.0
django-heroku==0.3.1
django-taggit==1.4.0
fastapi==0.61.1
Flask==2.0.0
gunicorn==20.1.0
h11==0.9.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pydantic==1.6.1
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
starlette==0.13.6
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.11.8
websockets==8.1
Werkzeug==2.0.1
whitenoise==5.2.0